move SQLQuery to it's own module
